Flytt funksjonelle tester for hentVilkårFor ut i egen fil
